refactor: Mode `fairings/` to `infrastrcture/middleware/`
