+++ api/src/application/mod.rs
@@ -1,4 +1,3 @@
-pub mod fairings;
 pub mod response;
 pub mod routes;
 pub mod services;
