+++ api/src/infrastructure/middleware/fairings/default_json.rs
