+++ api/src/infrastructure/mod.rs
@@ -1,6 +1,7 @@
 pub mod database;
 pub mod db_schema;
 pub mod logging;
+pub mod middleware;
 pub mod models;
 pub mod notify;
 pub mod paging;
