+++ api/src/lib.rs
@@ -11,9 +11,9 @@ use rocket::fs::FileServer;
 use rocket::{routes, Build, Config, Rocket};
 use rocket_db_pools::Database;
 
-use crate::application::fairings::{cors::CORS, default_json::DefaultJSON};
 use crate::application::routes::{health, jobs, monitors};
 use crate::infrastructure::database::{run_migrations, Db};
+use crate::infrastructure::middleware::fairings::{cors::CORS, default_json::DefaultJSON};
 
 #[rocket::launch]
 pub fn rocket() -> Rocket<Build> {
